+++ src/Console/SshCommand.php
@@ -81,7 +81,7 @@ class SshCommand extends Command
         }
 
         Context::push(new Context($host, $input, $output));
-        $options = $host->sshOptions();
+        $options = $host->getSshArguments();
         $deployPath = $host->get('deploy_path', '~');
 
         passthru("ssh -t $options $host 'cd '''$deployPath/current'''; exec \$SHELL -l'");